refactor: error handling to groupMemberExpenses

+++ backend/src/groups/expenses/sub-expenses/group-member-expenses/group-member-expenses.service.ts
@@ -1,25 +1,41 @@
 import { Injectable } from '@nestjs/common';
-import { CreateGroupMemberExpenseDto } from './dto/create-group-member-expense.dto';
+import {
+  CreateGroupMemberExpenseDto,
+  GROUP_MEMBER_EXPENSE_ROLE,
+} from './dto/create-group-member-expense.dto';
 import { UpdateGroupMemberExpenseDto } from './dto/update-group-member-expense.dto';
 import { PrismaClient } from '@prisma/client';
+import Big from 'big.js';
 const prisma = new PrismaClient();
 
+interface GroupMemberExpenseMany {
+  amount: string;
+  amountReferenceCurrency: string;
+  role: keyof typeof GROUP_MEMBER_EXPENSE_ROLE;
+  currency: string;
+  groupMemberId: string;
+  subExpenseId: string;
+}
 @Injectable()
 export class GroupMemberExpensesService {
-  create(
-    createGroupMemberExpenseDto: CreateGroupMemberExpenseDto,
+  async create(
+    createGroupMemberExpenseDto: CreateGroupMemberExpenseDto[],
     subExpenseId: string,
   ) {
-    return prisma.groupMemberExpense.create({
-      data: {
-        ...createGroupMemberExpenseDto,
-        amount: createGroupMemberExpenseDto.amount.toString(),
-        //TODO: Change 23 to an acutal thing
-        amountReferenceCurrency: 23,
-        groupMemberId: createGroupMemberExpenseDto.groupMemberId,
-        subExpenseId: subExpenseId,
-      },
+    const groupMemberExpenses: GroupMemberExpenseMany[] =
+      createGroupMemberExpenseDto.map((groupMemberExpense) => ({
+        ...groupMemberExpense,
+        amount: groupMemberExpense.amount.toString(),
+        //TODO: Needs to be changed from 23
+        amountReferenceCurrency: new Big(23).toString(),
+        subExpenseId,
+      }));
+
+    await prisma.groupMemberExpense.createMany({
+      data: groupMemberExpenses,
     });
+
+    return this.findAll(subExpenseId);
   }
 
   findAll(subExpenseId: string) {
@@ -76,11 +92,17 @@ export class GroupMemberExpensesService {
     });
   }
 
-  removeAllGroupMemberExpenseBelongingToSubExpense(subExpenseId: string) {
+  async removeAllGroupMemberExpensesBelongingToSubExpense(
+    subExpenseId: string,
+  ) {
+    const findResults = await this.findAll(subExpenseId);
+
+    if (findResults.length === 0) return;
+
     return prisma.groupMemberExpense.deleteMany({
       where: {
-        subExpenseId
-      }
-    })
+        subExpenseId,
+      },
+    });
   }
 }
